Conversion fixes (int vs. BYTE, signed vs. unsigned) and build system updates for the DDK compilers.
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@29 048eaad1-c872-4173-8c30-523bb29f53ec

+++ nibscan.c
@@ -411,7 +411,7 @@ scandisk(void)
 		if (track_length[track] > 0)
 		{
 			track_density[track] =
-			  check_sync_flags(track_buffer + (track * NIB_TRACK_LENGTH), track_density[track] & 3, track_length[track]);
+			  (BYTE)check_sync_flags(track_buffer + (track * NIB_TRACK_LENGTH), track_density[track] & 3, track_length[track]);
 
 			printf("(%d", track_density[track] & 3);
